ir: Fix redundant scalar allocation
