+++ devito/ir/iet/scheduler.py
@@ -1,6 +1,8 @@
 from collections import OrderedDict
 
-from devito.cgen_utils import Allocator
+import cgen as c
+
+from devito.cgen_utils import ccode
 from devito.ir.iet import (ArrayCast, Expression, Increment, LocalExpression, Element,
                            Iteration, List, Conditional, Section, HaloSpot,
                            ExpressionBundle, MapSections, Transformer, FindNodes,
@@ -103,69 +105,6 @@ def iet_lower_dimensions(iet):
     return iet
 
 
-def iet_insert_decls(iet, external):
-    """
-    Transform the input IET inserting the necessary symbol declarations.
-    Declarations are placed as close as possible to the first symbol occurrence.
-
-    Parameters
-    ----------
-    iet : Node
-        The input Iteration/Expression tree.
-    external : tuple, optional
-        The symbols defined in some outer Callable, which therefore must not
-        be re-defined.
-    """
-    iet = as_tuple(iet)
-
-    # Classify and then schedule declarations to stack/heap
-    allocator = Allocator()
-    mapper = OrderedDict()
-    for k, v in MapSections().visit(iet).items():
-        if k.is_Expression:
-            if k.is_scalar_assign:
-                # Inline declaration
-                mapper[k] = LocalExpression(**k.args)
-                continue
-            objs = [k.write]
-        elif k.is_Call:
-            objs = k.arguments
-
-        for i in objs:
-            try:
-                if i.is_LocalObject:
-                    # On the stack
-                    site = v if v else iet
-                    allocator.push_stack(site[-1], i)
-                elif i.is_Array:
-                    if i in as_tuple(external):
-                        # The Array is defined in some other IET
-                        continue
-                    elif i._mem_stack:
-                        # On the stack
-                        key = lambda i: not i.is_Parallel
-                        site = filter_iterations(v, key=key) or iet
-                        allocator.push_stack(site[-1], i)
-                    else:
-                        # On the heap, as a tensor that must be globally accessible
-                        allocator.push_heap(i)
-            except AttributeError:
-                # E.g., a generic SymPy expression
-                pass
-
-    # Introduce declarations on the stack
-    for k, v in allocator.onstack:
-        mapper[k] = tuple(Element(i) for i in v)
-    iet = Transformer(mapper, nested=True).visit(iet)
-
-    # Introduce declarations on the heap (if any)
-    if allocator.onheap:
-        decls, allocs, frees = zip(*allocator.onheap)
-        iet = List(header=decls + allocs, body=iet, footer=frees)
-
-    return iet
-
-
 def iet_insert_casts(iet, parameters):
     """
     Transform the input IET inserting the necessary type casts.
@@ -188,3 +127,130 @@ def iet_insert_casts(iet, parameters):
     casts = [ArrayCast(i) for i in parameters if i in need_cast]
     iet = List(body=casts + [iet])
     return iet
+
+
+def iet_insert_decls(iet, external):
+    """
+    Transform the input IET inserting the necessary symbol declarations.
+    Declarations are placed as close as possible to the first symbol occurrence.
+
+    Parameters
+    ----------
+    iet : Node
+        The input Iteration/Expression tree.
+    external : tuple, optional
+        The symbols defined in some outer Callable, which therefore must not
+        be re-defined.
+    """
+    iet = as_tuple(iet)
+
+    # Classify and then schedule declarations to stack/heap
+    allocator = Allocator()
+    for k, v in MapSections().visit(iet).items():
+        if k.is_Expression:
+            if k.is_scalar_assign:
+                # On the stack
+                site = v if v else iet
+                allocator.push_scalar_on_stack(site[-1], k)
+                continue
+            objs = [k.write]
+        elif k.is_Call:
+            objs = k.arguments
+
+        for i in objs:
+            try:
+                if i.is_LocalObject:
+                    # On the stack
+                    site = v if v else iet
+                    allocator.push_object_on_stack(site[-1], i)
+                elif i.is_Array:
+                    if i in as_tuple(external):
+                        # The Array is defined in some other IET
+                        continue
+                    elif i._mem_stack:
+                        # On the stack
+                        key = lambda i: not i.is_Parallel
+                        site = filter_iterations(v, key=key) or iet
+                        allocator.push_object_on_stack(site[-1], i)
+                    else:
+                        # On the heap
+                        allocator.push_array_on_heap(i)
+            except AttributeError:
+                # E.g., a generic SymPy expression
+                pass
+
+    # Introduce declarations on the stack
+    mapper = dict(allocator.onstack)
+    iet = Transformer(mapper, nested=True).visit(iet)
+
+    # Introduce declarations on the heap (if any)
+    if allocator.onheap:
+        decls, allocs, frees = zip(*allocator.onheap)
+        iet = List(header=decls + allocs, body=iet, footer=frees)
+
+    return iet
+
+
+class Allocator(object):
+
+    """
+    Support class to generate allocation code in different scopes.
+    """
+
+    def __init__(self):
+        self.heap = OrderedDict()
+        self.stack = OrderedDict()
+
+    def push_object_on_stack(self, scope, obj):
+        """Define an Array or a composite type (e.g., a struct) on the stack."""
+        handle = self.stack.setdefault(scope, OrderedDict())
+
+        if obj.is_LocalObject:
+            handle[obj] = Element(c.Value(obj._C_typename, obj.name))
+        else:
+            shape = "".join("[%s]" % ccode(i) for i in obj.symbolic_shape)
+            alignment = "__attribute__((aligned(%d)))" % obj._data_alignment
+            value = "%s%s %s" % (obj.name, shape, alignment)
+            handle[obj] = Element(c.POD(obj.dtype, value))
+
+    def push_scalar_on_stack(self, scope, expr):
+        """Define a Scalar on the stack."""
+        handle = self.stack.setdefault(scope, OrderedDict())
+
+        obj = expr.write
+        if obj in handle:
+            return
+
+        handle[obj] = None  # Placeholder to avoid reallocation
+        self.stack[expr] = LocalExpression(**expr.args)
+
+    def push_array_on_heap(self, obj):
+        """Define an Array on the heap."""
+        if obj in self.heap:
+            return
+
+        decl = "(*%s)%s" % (obj.name, "".join("[%s]" % i for i in obj.symbolic_shape[1:]))
+        decl = c.Value(obj._C_typedata, decl)
+
+        shape = "".join("[%s]" % i for i in obj.symbolic_shape)
+        alloc = "posix_memalign((void**)&%s, %d, sizeof(%s%s))"
+        alloc = alloc % (obj.name, obj._data_alignment, obj._C_typedata, shape)
+        alloc = c.Statement(alloc)
+
+        free = c.Statement('free(%s)' % obj.name)
+
+        self.heap[obj] = (decl, alloc, free)
+
+    @property
+    def onstack(self):
+        ret = []
+        for k, v in self.stack.items():
+            try:
+                ret.append((k, [i for i in v.values() if i is not None]))
+            except AttributeError:
+                ret.append((k, v))
+        return ret
+
+    @property
+    def onheap(self):
+        return self.heap.values()
